test: split PDF text operation attestation coverage
